updated definition -- using scripts
