+++ declarations/jersey/index.d.ts
@@ -1,5 +1,5 @@
 import { ColorValue, ViewStyle } from 'react-native';
-import SleeperJersey from 'components/ui/jersey';
+import SleeperJersey from '@sleeperhq/mini-core/declarations/types/components/jersey.d';
 export interface JerseyProps {
     style: ViewStyle;
     sport: 'nfl' | 'nba' | 'cbb' | 'cfb' | 'mlb';
