+++ declarations/switch/index.d.ts
@@ -1,3 +1,3 @@
-import { AppIconSwitch } from '../types/components/ui_2/app_icon_switch';
-export type { SwitchOption, Props as SwitchProps } from '../types/components/ui_2/app_icon_switch';
+import { AppIconSwitch } from '@sleeperhq/mini-core/declarations/types/components/app_icon_switch.d';
+export type { SwitchOption, Props as SwitchProps } from '@sleeperhq/mini-core/declarations/types/components/app_icon_switch.d';
 export default AppIconSwitch;
