+++ declarations/types/components/jersey.d.ts
@@ -1,0 +1,3 @@
+export default Jersey;
+declare const Jersey: React.NamedExoticComponent<object>;
+import React from 'react';
