+++ declarations/types/shared/graphql.d.ts
@@ -202,6 +202,35 @@ export type RosterDraftPick = {
     round?: Maybe<Scalars['Int']>;
     season?: Maybe<Scalars['String']>;
 };
+export type Topic = {
+    __typename?: 'Topic';
+    attachment?: Maybe<Scalars['Json']>;
+    author_avatar?: Maybe<Scalars['String']>;
+    author_display_name?: Maybe<Scalars['String']>;
+    author_id?: Maybe<Scalars['Snowflake']>;
+    author_is_bot?: Maybe<Scalars['Boolean']>;
+    channel_id?: Maybe<Scalars['Snowflake']>;
+    channel_tags?: Maybe<Scalars['Set']>;
+    client_id?: Maybe<Scalars['String']>;
+    created?: Maybe<Scalars['Int']>;
+    hidden?: Maybe<Scalars['Boolean']>;
+    last_message_id?: Maybe<Scalars['Snowflake']>;
+    last_pinned_message_id?: Maybe<Scalars['Snowflake']>;
+    last_read_id?: Maybe<Scalars['Snowflake']>;
+    num_messages?: Maybe<Scalars['Int']>;
+    num_viewers?: Maybe<Scalars['Int']>;
+    pinned?: Maybe<Scalars['Boolean']>;
+    player_tags?: Maybe<Scalars['Set']>;
+    pushed_by?: Maybe<Scalars['Json']>;
+    score?: Maybe<Scalars['Int']>;
+    shadowed?: Maybe<Scalars['Boolean']>;
+    shard_max?: Maybe<Scalars['Int']>;
+    shard_min?: Maybe<Scalars['Int']>;
+    title?: Maybe<Scalars['String']>;
+    title_map?: Maybe<Scalars['Json']>;
+    topic_id?: Maybe<Scalars['Snowflake']>;
+    upvotes?: Maybe<Scalars['Int']>;
+};
 export type User = {
     __typename?: 'User';
     async_bundles?: Maybe<Scalars['List']>;
@@ -226,32 +255,4 @@ export type User = {
     user_id?: Maybe<Scalars['Snowflake']>;
     username?: Maybe<Scalars['String']>;
     verification?: Maybe<Scalars['String']>;
-};
-export type Topic = {
-  __typename?: 'Topic';
-  attachment?: Maybe<Scalars['Json']>;
-  author_avatar?: Maybe<Scalars['String']>;
-  author_display_name?: Maybe<Scalars['String']>;
-  author_id?: Maybe<Scalars['Snowflake']>;
-  author_is_bot?: Maybe<Scalars['Boolean']>;
-  channel_id?: Maybe<Scalars['Snowflake']>;
-  channel_tags?: Maybe<Scalars['Set']>;
-  client_id?: Maybe<Scalars['String']>;
-  created?: Maybe<Scalars['Int']>;
-  hidden?: Maybe<Scalars['Boolean']>;
-  last_message_id?: Maybe<Scalars['Snowflake']>;
-  last_pinned_message_id?: Maybe<Scalars['Snowflake']>;
-  last_read_id?: Maybe<Scalars['Snowflake']>;
-  num_messages?: Maybe<Scalars['Int']>;
-  num_viewers?: Maybe<Scalars['Int']>;
-  pinned?: Maybe<Scalars['Boolean']>;
-  player_tags?: Maybe<Scalars['Set']>;
-  pushed_by?: Maybe<Scalars['Json']>;
-  score?: Maybe<Scalars['Int']>;
-  shard_max?: Maybe<Scalars['Int']>;
-  shard_min?: Maybe<Scalars['Int']>;
-  title?: Maybe<Scalars['String']>;
-  title_map?: Maybe<Scalars['Json']>;
-  topic_id?: Maybe<Scalars['Snowflake']>;
-  upvotes?: Maybe<Scalars['Int']>;
 };
