+++ declarations/types/utils/toast_helper.d.ts
@@ -1,7 +1,7 @@
 import { TextStyle, ViewStyle } from 'react-native';
 export type ToastConfig = {
     text: string;
-    icon?: string;
+    icon?: 'success' | 'error' | 'trade' | 'voice';
     time?: number;
     style?: ViewStyle;
     slideTop?: boolean;
